+++ R/methods.R
@@ -8,7 +8,7 @@ mround <- function(x, digits) {
   sprintf(paste0("%.", digits, "f"), round(x, digits))
 }
 
-#' Extract the number of observations of a sdmTMB model
+#' Extract the number of observations of an sdmTMB model
 #'
 #' @param object The fitted sdmTMB model object
 #' @importFrom stats nobs
@@ -16,6 +16,15 @@ mround <- function(x, digits) {
 #' @noRd
 nobs.sdmTMB <- function(object, ...) {
     sum(!is.na(object$data[all.vars(object$formula[[1]])[1]]))
+}
+
+#' Get fitted values from an sdmTMB model
+#'
+#' @param object The fitted sdmTMB model object
+#' @importFrom stats predict
+#' @export
+fitted.sdmTMB <- function(object, ...) {
+  predict(object, type = "response")$est
 }
 
 #' Extract the log likelihood of a sdmTMB model